requirements.txt
matplotlib
pytorch-ignite
torchvision
torchviz
torch
scipy
pynvml
tqdm
decorator
kornia
sklearn
h5py
Pillow-SIMD
pytorch-model-summary
